# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/internal/admin/payouts" : {
      "get" : {
        "tags" : [ "payout-controller" ],
        "operationId" : "getPayouts_2",
        "parameters" : [ {
          "name" : "payeeType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "payeeId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoPayoutDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "payout-controller" ],
        "operationId" : "createPayout",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePayoutRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Payout"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/cash/remit" : {
      "post" : {
        "tags" : [ "cash-controller" ],
        "operationId" : "remitCash",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CashRemittanceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashRemittance"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ledger/payouts/{payeeType}/{payeeId}" : {
      "get" : {
        "tags" : [ "payee-payout-controller" ],
        "operationId" : "getPayouts",
        "parameters" : [ {
          "name" : "payeeType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "payeeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoPayoutDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ledger/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "payee-cash-controller" ],
        "operationId" : "getCashByDriver",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoCashRemittanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "payee-cash-controller" ],
        "operationId" : "getCashSummary",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/ledger/payouts" : {
      "get" : {
        "tags" : [ "internal-payout-controller" ],
        "operationId" : "getPayouts_1",
        "parameters" : [ {
          "name" : "payeeType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "payeeId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoPayoutDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/ledger/payouts/latest/{payeeType}/{payeeId}" : {
      "get" : {
        "tags" : [ "internal-payout-controller" ],
        "operationId" : "getPayeeSummary",
        "parameters" : [ {
          "name" : "payeeType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "payeeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PayeeMoneySummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/ledger/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "internal-cash-controller" ],
        "operationId" : "getCashByDriver_1",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoCashRemittanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/ledger/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "internal-cash-controller" ],
        "operationId" : "getCashSummary_1",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "cash-controller" ],
        "operationId" : "getCashByDriver_2",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageCashRemittance"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "cash-controller" ],
        "operationId" : "getCashSummary_2",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreatePayoutRequest" : {
        "type" : "object",
        "properties" : {
          "payeeType" : {
            "type" : "string"
          },
          "payeeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "periodTo" : {
            "type" : "string",
            "format" : "date-time"
          },
          "force" : {
            "type" : "boolean"
          }
        }
      },
      "Payout" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "payeeType" : {
            "type" : "string"
          },
          "payeeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "payeeDisplayName" : {
            "type" : "string"
          },
          "periodFrom" : {
            "type" : "string",
            "format" : "date-time"
          },
          "periodTo" : {
            "type" : "string",
            "format" : "date-time"
          },
          "amount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "APPROVED", "PAID", "FAILED", "CANCELLED" ]
          },
          "beneficiarySnapshot" : {
            "type" : "string"
          },
          "bankReference" : {
            "type" : "string"
          },
          "failureReason" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "approvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "paidBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "settledTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "approvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "paidAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CashRemittanceRequest" : {
        "type" : "object",
        "properties" : {
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "reference" : {
            "type" : "string"
          }
        }
      },
      "CashRemittance" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "reference" : {
            "type" : "string"
          },
          "recordedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
      "PageResponseDtoPayoutDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PayoutDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PayoutDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "payeeType" : {
            "type" : "string"
          },
          "payeeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "payeeDisplayName" : {
            "type" : "string"
          },
          "periodFrom" : {
            "type" : "string",
            "format" : "date-time"
          },
          "periodTo" : {
            "type" : "string",
            "format" : "date-time"
          },
          "amount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "beneficiarySnapshot" : {
            "type" : "string"
          },
          "bankReference" : {
            "type" : "string"
          },
          "failureReason" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "approvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "paidBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "settledTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "approvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "paidAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CashRemittanceDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "reference" : {
            "type" : "string"
          },
          "recordedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoCashRemittanceDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CashRemittanceDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "CashSummaryDto" : {
        "type" : "object",
        "properties" : {
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "cashCollected" : {
            "type" : "number"
          },
          "cashRemitted" : {
            "type" : "number"
          },
          "cashInHand" : {
            "type" : "number"
          }
        }
      },
...
      "BeneficiaryResponse" : {
        "type" : "object",
        "properties" : {
          "accountNumberMasked" : {
            "type" : "string"
          },
          "ifsc" : {
            "type" : "string"
          },
          "beneficiaryName" : {
            "type" : "string"
          },
          "verified" : {
            "type" : "boolean"
          },
          "source" : {
            "type" : "string"
          }
        }
      },
      "PayeeMoneySummaryDto" : {
        "type" : "object",
        "properties" : {
          "payeeType" : {
            "type" : "string"
          },
          "payeeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "unsettledAmount" : {
            "type" : "number"
          },
          "pendingPayoutAmount" : {
            "type" : "number"
          },
          "lastPayout" : {
            "$ref" : "#/components/schemas/PayoutDto"
          },
          "beneficiary" : {
            "$ref" : "#/components/schemas/BeneficiaryResponse"
          }
        }
      },
...
      "PageCashRemittance" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CashRemittance"
            }
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}